+++ src/lib/components/Button.example.border.js
@@ -1,0 +1,8 @@
+var React = require("react");
+var Button = require("./Button");
+
+/*eslint no-unused-vars: 0*/
+
+module.exports = (
+	<Button width={250} height={40} bgColor="#ff00ff" borderColor="#ff0000" href="http://google.com/">Learn more</Button>
+);