+++ src/lib/components/Divider.example.js
@@ -1,0 +1,8 @@
+var React = require("react");
+var Divider = require("./Divider");
+
+/*eslint no-unused-vars: 0*/
+
+module.exports = (
+	<Divider color="#000000" thickness={2} />
+);