+++ src/lib/components/Hero.example.js
@@ -1,0 +1,10 @@
+var React = require("react");
+var Hero = require("./Hero");
+
+/*eslint no-unused-vars: 0*/
+
+module.exports = (
+	<Hero width={600} height={400} background="images/hero.jpg" bgColor="#ffffff">
+		This is in the middle of the hero.
+	</Hero>
+);