+++ src/lib/components/Image.example.js
@@ -1,0 +1,8 @@
+var React = require("react");
+var Image = require("./Image");
+
+/*eslint no-unused-vars: 0*/
+
+module.exports = (
+	<Image width="200" src="images/logo.jpg" />
+);