+++ src/lib/components/OrderedList.example.js
@@ -1,0 +1,12 @@
+var React = require("react");
+var OrderedList = require("./OrderedList");
+var ListItem = require("./ListItem");
+
+/*eslint no-unused-vars: 0*/
+
+module.exports = (
+	<OrderedList>
+		<ListItem>Item <strong>one</strong>.</ListItem>
+		<ListItem>Item two.</ListItem>
+	</OrderedList>
+);