+++ src/lib/components/ParaBlock.example.js
@@ -1,0 +1,12 @@
+var React = require("react");
+var ParaBlock = require("./ParaBlock");
+var Para = require("./Para");
+
+/*eslint no-unused-vars: 0*/
+
+module.exports = (
+	<ParaBlock>
+		<Para>Paragraph <strong>one</strong>.</Para>
+		<Para>Paragraph two.</Para>
+	</ParaBlock>
+);