+++ src/lib/components/RawHtml.example.js
@@ -1,0 +1,8 @@
+var React = require("react");
+var RawHtml = require("./RawHtml");
+
+/*eslint no-unused-vars: 0*/
+
+module.exports = (
+	<RawHtml>{"<!-- This will be redered as an HTML comment -->"}</RawHtml>
+);