+++ src/lib/components/layout/Col.example.js
@@ -1,0 +1,8 @@
+var React = require("react");
+var Col = require("./Col");
+
+/*eslint no-unused-vars: 0*/
+
+module.exports = (
+	<Col wrapper={{align: "center"}}>My Column</Col>
+);