+++ src/lib/components/layout/Container.example.js
@@ -1,0 +1,10 @@
+var React = require("react");
+var Container = require("./Container");
+
+/*eslint no-unused-vars: 0*/
+
+module.exports = (
+	<Container>
+		My Content
+	</Container>
+);