+++ src/lib/components/layout/Frame.example.js
@@ -1,0 +1,10 @@
+var React = require("react");
+var Frame = require("./Frame");
+
+/*eslint no-unused-vars: 0*/
+
+module.exports = (
+	<Frame borderColor="#ff0000">
+		My Content
+	</Frame>
+);