+++ src/lib/components/layout/FullWidthRow.example.js
@@ -1,0 +1,12 @@
+var React = require("react");
+var FullWidthRow = require("./FullWidthRow");
+var Col = require("./Col");
+
+/*eslint no-unused-vars: 0*/
+
+module.exports = (
+	<FullWidthRow bgColor="#eaeaea">
+		<Col>One</Col>
+		<Col>Two</Col>
+	</FullWidthRow>
+);