+++ src/lib/components/layout/Table.example.js
@@ -1,0 +1,15 @@
+var React = require("react");
+var Table = require("./Table");
+
+/*eslint no-unused-vars: 0*/
+
+module.exports = (
+	<Table>
+		<tbody>
+			<tr>
+				<td>Col 1</td>
+				<td>Col 2</td>
+			</tr>
+		</tbody>
+	</Table>
+);